moving the dropdown component from viral
