+++ app/assets/stylesheets/application.tailwind.css
@@ -105,12 +105,12 @@ input[type="search"] {
 @utility navbar-button {
   @apply flex min-h-11 min-w-11 items-center justify-center rounded-lg p-1 text-sm text-slate-500 hover:bg-slate-300 dark:text-slate-400 dark:hover:bg-slate-700;
 
-  & button.viral-dropdown--icon {
+  & button.dropdown--icon {
     @apply rounded-lg p-1 text-sm text-slate-500 hover:bg-slate-300 dark:text-slate-400 dark:hover:bg-slate-700 dark:hover:text-white;
   }
 }
 
-@utility viral-dropdown--icon {
+@utility dropdown--icon {
   & button.navbar-button {
     @apply rounded-lg p-1 text-sm text-slate-500 hover:bg-slate-300 dark:text-slate-400 dark:hover:bg-slate-700;
   }
@@ -392,7 +392,7 @@ input[type="search"] {
   @apply text-primary-500;
 }
 
-@utility viral-dropdown--button {
+@utility dropdown--button {
   @apply hover:text-primary-700 flex min-h-11 min-w-11 items-center rounded-lg border border-slate-200 bg-white px-5 py-2.5 text-sm font-medium text-slate-900 hover:bg-slate-100 focus-visible:z-10 dark:border-slate-600 dark:bg-slate-800 dark:text-slate-300 dark:hover:bg-slate-700 dark:hover:text-white;
 }
