+++ test/components/previews/dropdown_component_preview.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-class ViralDropdownComponentPreview < ViewComponent::Preview
+class DropdownComponentPreview < ViewComponent::Preview
   def default
-    render Viral::DropdownComponent.new(label: 'Organism', aria: { label: 'Organism dropdown list' },
-                                        title: 'Organisms that really shine') do |dropdown|
+    render DropdownComponent.new(label: 'Organism', aria: { label: 'Organism dropdown list' },
+                                 title: 'Organisms that really shine') do |dropdown|
       dropdown.with_item(label: 'Aspergillus awamori', url: '#')
       dropdown.with_item(label: 'Bacillus cereus', url: '#')
       dropdown.with_item(label: 'Pseudomonas aeruginosa', url: '#')
@@ -11,8 +11,8 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
   end
 
   def with_caret
-    render Viral::DropdownComponent.new(label: 'Organism', aria: { label: 'Organism dropdown list' },
-                                        caret: true) do |dropdown|
+    render DropdownComponent.new(label: 'Organism', aria: { label: 'Organism dropdown list' },
+                                 caret: true) do |dropdown|
       dropdown.with_item(label: 'Aspergillus awamori', url: '#')
       dropdown.with_item(label: 'Bacillus cereus', url: '#')
       dropdown.with_item(label: 'Pseudomonas aeruginosa', url: '#')
@@ -20,8 +20,8 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
   end
 
   def with_icon
-    render Viral::DropdownComponent.new(icon: :plus_circle,
-                                        aria: { label: 'Organism dropdown list' }) do |dropdown|
+    render DropdownComponent.new(icon: :plus_circle,
+                                 aria: { label: 'Organism dropdown list' }) do |dropdown|
       dropdown.with_item(label: 'Aspergillus awamori', url: '#')
       dropdown.with_item(label: 'Bacillus cereus', url: '#')
       dropdown.with_item(label: 'Pseudomonas aeruginosa', url: '#')
@@ -29,8 +29,8 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
   end
 
   def with_item_icon
-    render Viral::DropdownComponent.new(icon: :list,
-                                        aria: { label: 'Organism Dropdown Menu' }) do |dropdown|
+    render DropdownComponent.new(icon: :list,
+                                 aria: { label: 'Organism Dropdown Menu' }) do |dropdown|
       dropdown.with_item(label: 'Checkmark', url: '#', icon_name: :check)
       dropdown.with_item(label: 'Inbox', url: '#', icon_name: :inbox_stack, icon_library: :heroicons)
     end
@@ -38,7 +38,7 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
 
   # Preview for custom button_styles
   def with_custom_button_styles
-    render(Viral::DropdownComponent.new(
+    render(DropdownComponent.new(
              label: 'Custom Button',
              styles: { button: 'bg-emerald-800 text-white px-4 py-2 flex gap-2 rounded-full cursor-pointer' },
              caret: true
@@ -51,7 +51,7 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
 
   # Preview for dropdown with tooltip
   def with_tooltip
-    render(Viral::DropdownComponent.new(
+    render(DropdownComponent.new(
              label: 'Tooltip Button',
              tooltip: 'This is a tooltip!',
              tooltip_placement: :bottom,
@@ -62,7 +62,7 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
   end
 
   def with_icon_and_tooltip
-    render Viral::DropdownComponent.new(
+    render DropdownComponent.new(
       icon: :list,
       tooltip: 'This is a tooltip!',
       tooltip_placement: :bottom,
@@ -74,7 +74,7 @@ class ViralDropdownComponentPreview < ViewComponent::Preview
 
   # Preview for dropdown items with data attributes
   def with_data_attributes
-    render Viral::DropdownComponent.new(
+    render DropdownComponent.new(
       label: 'Data Attributes Test',
       caret: true
     ) do |dropdown|
